ore/gcloud: Add --write-url argument
coreos-assembler was trying to re-create the URL that ore computed
for the storage, and a recent refactoring broke that.

Add an option to write out the URL we used so we only need
to have its computation in one place.

Specifically, cosa started replacing `.` with `-` but ore wasn't.

xref https://github.com/coreos/coreos-assembler/issues/1078