qemuvariants: simplify VMware OVA generation by using callback
cosalib.vmware overrode mutate_image and then called
QemuVariant.mutate_image with the callback. While working a seperate
issue, I started to define another class to use the call-back mechanism.
Rather than introduce another Class, this change allows for a class
attribute of "mutate_callback" to be defined which is used by default
unless `mutate_image(callback=foo)` is called.

The advantage to this approach is that VMware's OVA can be a bit
simpler.

Finally, a format check was added to ensure that the format is
appropriately changed. In some cases, 'qemu-img convert' will produce
'raw' images instead of compliant VHD's.