upload-oscontainer: New command
This code lives today in the RHCOS pipeline (internal gitlab);
to be honest, I have no idea why only the "core" oscontainer bits
ended up in cosa upstream and the script lived downstream.

Let's fix that, moving all of the logic here.  Prep for
further work.

Make this work a bit better too by automatically copying in
`~/.docker/config.json` into the supermin appliance, which is
something the RHCOS pipeline does.

And yes I hope to clean this all up more in the future but
for now let's get the code out of the darkness and into the light.