kola: Remove fcos.python test
Let's rely on https://github.com/coreos/fedora-coreos-config/pull/280 -
starting a whole VM for this is a bit silly.  And in the future
if we do add more tests like this I think we should do it via
tests that run the OS as a container, or are part of a set of
explicitly "nondestructive" tests that we can run multiple in a
single VM with less overhead etc.