kola: Remove SimpleEtcd
While working on a PR related to Ignition conversion I came
across this special `$discovery` bit.  Since etcd is not
a core feature of FCOS by default and instead something
that can be enabled in a containerized fashion, let's
remove the `SimpleEtcd` code.  Other projects that want
etcd can carry code to set it up.  Or we can re-add it
at a higher level here in kola again rather than wedged
in as part of the platform setup.