Drop ostree-releng-scripts git submodule
While we're in the mood for deleting git submodules, let's also delete
the ostree-releng-scripts. In practice, we don't use it in cosa itself,
and it's trivial for anyone who wants a script from there while e.g.
debugging from a cosa container to `git clone` it.

This also allows us to clean up the build script a bit.