mantle/*: Add SilenceUsage to all CLI entrypoints
The cobra library made so many other good decisions, like
literally calling it cobra for commands which always summons
memories of GI Joe from my childhood.

This decision to print usage when an error is thrown is not
one of those good decisions; it spams output unnecessarily.

And holy cow do we have a lot of commmands.

Anyways, add `SilenceUsage` to all of them.