mantle/system: Add RpmArch API
A lot of our code uses the portage architecture, which no longer
makes sense.  Now, the more correct code for this is part of
rpm-ostree but I'm not sure we want to go to forking off that yet.
Since the build system only supports this set of architectures
currently, let's just inline the Go -> RPM translation here.