cmd-build: Add hack for composing across refs
I spend a lot of my time building different branches of
fedora-coreos-config to debug issues. One problem is that when changing
refs like this, there's no guarantee that rpm-ostree won't re-use the
same version string. If that happens, the build will fail with `EEXIST`
at the end when trying to move the builddir in its final place.

Work around this corner-case by pointing the new ref at the old ref. As
mentioned in the comment, this is a crude not-technically-correct hack
which in practice shouldn't really break anything. We also force a
rebuild (which should happen anyway since at least the commit metadata
changed) to be sure the ref is updated.