metal/testiso: Don't inherit console by default, log cp errors
The console is useful for debugging but in the success path
is way too verbose.  What we really want to do is hook into
the console the same way main kola tests do.  That will
come in the future.

Also log errors from `cp` since that's happening in CI (probably
`ENOSPC`); and try to fix that by setting `TMPDIR` in CI.