mantle: Rework testiso/metal to use new qemu channel API
First, add an API for creating channels that give a pipe
to the Go side, so we don't need to write to temp files that
we then read.

Now, I'm still trying to fine-tune the interface between the
"metal" API and the more hardcoded "testiso".

Tweak things so that the install path supports being provided
a builder object, but add an API that creates a default one
that mostly just sets up the default max memory.

(Also, I stopped trying to special case a lower memory for
 the legacy install because that's going to go away anyways)