Kola: add support for multipath devices
Add support for "faking" multipath devices. In order to do this, Kola
converts QCOW2 to a raw device and adds the device twice with attributes
sufficient to trick multipath tools into seeing the devices as two paths
to the same device.

This feature is needed in order to adequately test root on multipath
devices for RHCOS.