qemuexec: Support --ignition-direct
I'm trying to test an Ignition config doing partitioning for RHCOS
and the model of trying to handle everything as spec 3 internally
and convert isn't going to work.  (Or at least, I don't want
to try to debug conversion code at the same time as this)

Add an option to say "just pass this config to the instance directly".