kola: fall back to upgrade testing from testing-devel for new streams
A new stream doesn't have any releases yet, so we can't do upgrade
testing within that stream. Instead, let's just fallback to upgrade
testing from testing-devel, which should always work since in theory all
the new streams we will be introducing are "ahead" in content from
testing-devel.

This is also a first step towards supporting cross-stream upgrade
testing. The next commit will add enablement for testing this in
`fcos.upgrade.basic`. The follow-up to this then is simply exposing the
base stream to use on the CLI).