fcos.upgrade.basic: support rebasing to different stream
We allow nodes to switch streams using `rpm-ostree rebase`. Let's make
our upgrade test reflect that. One cool thing to note is that the
`upgrade-from-current` test does still use Zincati, so this still does
some sanity-checking of Zincati on the new stream.

This goes together with the previous commit, in which we've added some
support for selecting a separate starting stream.