qemu: Lazily convert passed Ignition config
Today for architectures without `-fw-cfg` we end up rendering
the Ignition config when `SetPrimaryDisk()` is called.  This
is a confusing trap because it forces us to call `SetConfig()`
before that.

Rework the API to render the config lazily, which doesn't
fix the problem yet - but we can do so by then delaying
injecting the config to `Exec()` time which I'll do as a followup.

Also while we have the patient open, use the new
`conf.SerializeAndMaybeConvert()` API.