devshell: Print serial console as status line, dump when interrupted
This is a followup to the previous progress status for
https://github.com/coreos/coreos-assembler/issues/1367

Rather than trying to parse the serial console, just print it as
is, except strip control characters/newlines so that it becomes
effectively a progress bar.  This way, we don't corrupt the terminal,
and we don't fill up its history either.

And, go a step farther and dump the whole serial console in the
case where we get Ctrl-C before a successful login.  This aids
in debugging.