kola/testiso: add support for offline testing
Add a new `iso-offline-install` scenario which tests the osmet path. In
this path, we completely turn off networking to the host (`-nic none`).

We pass `--offline` to force coreos-installer to use the osmet path. In
the future, we should be able to omit this entirely, which will test
that coreos-installer correctly defaults to osmet when not provided
another image source.

As mentioned in the comment, this might become the default behaviour of
`iso-install` in the future once osmet is fully baked in.