buildextend-live: turn on osmet packing by default
This is the final step in ratcheting osmet into place for offline
installs by default. We drop the `--osmet` optional flag and now always
perform osmet packing when generating the live ISO.

(We could leave the switch around for a while just for compatibility,
but meh... the only other place this exists right now is the
coreos-installer CI, where we can trivially update it once this merges.)

Though because RHCOS doesn't yet support osmet, we add support for an
`osmet` knob in `image.yaml` which the RHCOS pipeline can use to turn it
off for now.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/352